split each implementation from constructor
